+++ src/main/java/com/smunity/api/domain/question/repository/AnswerRepository.java
@@ -1,6 +1,6 @@
-package com.smunity.api.domain.qna.repository;
+package com.smunity.api.domain.question.repository;
 
-import com.smunity.api.domain.qna.domain.Answer;
+import com.smunity.api.domain.question.domain.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
