+++ src/main/java/com/smunity/api/domain/question/service/QuestionService.java
@@ -1,6 +1,6 @@
-package com.smunity.api.domain.qna.service;
+package com.smunity.api.domain.question.service;
 
-import com.smunity.api.domain.qna.dto.QuestionDto;
+import com.smunity.api.domain.question.dto.QuestionDto;
 import java.util.List;
